crypto: mbedtls_aes_alt: add API implementation, input validation, fix initialization and data types
1. add alt implementation mbedtls_aes_setkey_dec() to complete mbedtls dependency.

2. The mbedtls_cipher_update() in mbedtls invokes the ecb_func
function (via function pointer) and expects 0 when encryption
is successful. Returning the the length resutls in wrong interpretation
of the result. Fix this.

3. The mbedtls_aes_context's keylen member should be unsigned int
to be consistent with the datatype passed in, or else this results
in overflow for AES_256_KEY_LEN (32).

4. Add checks to validate input.

+++ hw/drivers/crypto/include/crypto/aes_alt.h
@@ -30,11 +30,10 @@ extern "C" {
 
 #include <crypto/crypto.h>
 
-typedef struct mbedtls_aes_context
-{
+typedef struct mbedtls_aes_context {
     struct crypto_dev *crypto;
     uint8_t key[AES_MAX_KEY_LEN];
-    uint8_t keylen;
+    unsigned int keylen;
 } mbedtls_aes_context;
 
 void mbedtls_aes_init(mbedtls_aes_context *ctx);
